Updated the export
he "Export" option is defined as a Dropdown.Item element. Clicking on this element sets the showExportModal state to true, which shows the export modal. The modal contains a form where the user can select the format they want to export the note in.

Clicking on the "Export" button in the modal calls the handleNoteExport function with the noteId and format as arguments, which initiates the export process.