Copy Note
Note that the copyNote function sends a POST request to the backend API to create a new note with the same content as the original note, but with a different title and category. The copiedFrom field is added to indicate that the new note is a copy of the original note.